Move user state tests to pytest
