+++ tests/filename_map.yml
@@ -16,7 +16,7 @@ salt/modules/(apkpkg|aptpkg|ebuildpkg|dpkg_lowpkg|freebsdpkg|mac_brew_pkg|mac_po
   - integration.states.test_pkgrepo
 
 salt/modules/(mac_user|useradd|pw_user|solaris_user|win_useradd)\.py:
-  - unit.states.test_user
+  - pytests.unit.states.test_user
   - integration.states.test_user
 
 salt/modules/(aix_group|groupadd|mac_group|pw_group|solaris_group|win_groupadd)\.py:
